requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
lxml==4.5.2
multidict==4.7.6
mysql-connector-python==8.0.21
protobuf==3.12.2
PyExecJS==1.5.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
typing-extensions==3.7.4.2
urllib3==1.25.10
yarl==1.5.0
